+++ src/pymyinstall/fix/__init__.py
@@ -4,3 +4,4 @@
 """
 
 from .releases import fix_scipy10_for_statsmodels08
+from .win32py37 import fix_win32ctypes_core_cffi__advapi32_py